requirements.txt
boto3==1.11.4
botocore==1.14.4
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0.13
django-chartit==0.2.9
django-disqus==0.5
django-highchartit==0.2.3
django-mathfilters==0.4.0
django-storages==1.8
django-widget-tweaks==1.4.3
docutils==0.15.2
fanstatic==1.1
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
js.respond==1.2.0
mock==3.0.5
mysqlclient==1.3.12
Pillow==5.4.1
pycparser==2.19
PyMySQL==0.9.3
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.1
requests==2.21.0
s3transfer==0.3.1
shutilwhich==1.1.0
simplejson==3.16.0
six==1.12.0
split==0.4
stripe==2.25.0
tinycss2==0.6.1
urllib3==1.24.2
WeasyPrint==44
webencodings==0.5.1
WebOb==1.8.5
